Load only the plugins for the component's of the greatest versions. Task 63080
Explicitly load some facets (i.e.: "components" and
"required_component") as part of cabar core, since they are required
for locating components in the Loader and resolving components in the
Resolver.

Notify Loaders of :plugin_installed, even if the plugin file has already
been loaded.

This took much longer than expectedr; the loading of the latest plugins
.vs. only the plugins of the activated components is probably a hack,
but it's probably the best solution for the short-term.